Fuse attributes when possible
